Made profile cards along with Grid to hold all rendered cards. Currently works with mock user data stored in mockData.js.
