+++ src/data/mockData.js
@@ -17,20 +17,11 @@ export default [
     "Back-End Tech-Stack": [
       "Linux",
       "Apache",
-      "MySQL",
-      "PHP",
       "MongoDB",
       "Express.JS",
-      "AngularJS",
       "Node.js",
-      "Javascript",
-      "Ruby",
       "SQLite",
-      "Rails",
       "Python",
-      "Linux",
-      "Nginx",
-      "Java",
     ],
     specialty: "Full-Stack",
     "Preferred Tech-Stack": ["HTML", "CSS", "Javascript", "Angular", "Java"],
@@ -77,14 +68,7 @@ export default [
     state: "NY ",
     "Preferred Coding Time": ["morning"],
     "Front-End Tech-Stack": ["HTML", "CSS", "Vue"],
-    "Back-End Tech-Stack": [
-      "Linux",
-      "SQLite",
-      "Rails",
-      "Python",
-      "Linux",
-      "Nginx",
-    ],
+    "Back-End Tech-Stack": ["Linux", "SQLite", "Rails", "Python", "Nginx"],
     specialty: "Front-End",
     "Preferred Tech-Stack": ["HTML", "CSS", "Javascript", "Vue"],
     bio:
@@ -122,7 +106,7 @@ export default [
     state: "IN",
     "Preferred Coding Time": ["evening"],
     "Front-End Tech-Stack": ["Javascript", "Angular", "Vue"],
-    "Back-End Tech-Stack": ["Linux", "Rails", "Python", "Linux", "Nginx"],
+    "Back-End Tech-Stack": ["Linux", "Rails", "Python", "Nginx"],
     specialty: "Full-Stack",
     "Preferred Tech-Stack": [
       "Apache",
@@ -145,7 +129,7 @@ export default [
     state: "IL",
     "Preferred Coding Time": ["afternoon"],
     "Front-End Tech-Stack": ["Javascript"],
-    "Back-End Tech-Stack": ["Linux", "Python", "Linux", "Nginx"],
+    "Back-End Tech-Stack": ["Python", "Linux", "Nginx"],
     specialty: "Front-End",
     "Preferred Tech-Stack": [
       "HTML",
